+++ RELEASE_NOTES.md
@@ -5,6 +5,8 @@
 - Added padding rules to HorizontalLayoutManager
 - Added spacing rules to HorizontalLayoutManager
 - Added documentation to LayoutManager
+- Added system-wide event injection for PushrodCallbackEvents (#155)
+
 
 ## 0.3.5